requirements.txt
Django==1.11
django-fieldbook
fieldbook_py
django-material
#
# Heroku
#
python-decouple
gunicorn==19.6.0
psycopg2==2.6.2
dj-database-url==0.4.1
#
# END - heroku dependecies
#